Add files via upload
We are conducting an A/B Test to check if a new webpage design for a hypothetical company is better than
the previous version in terms of conversion rate (ie. did the user make a purchase on the webpage).

Control Group : Users who are exposed to the old webpage design
Test Group: Users who are exposed to the new webpage design.